mpi: switched to floats

+++ mpi/src/data.c
@@ -5,7 +5,7 @@
 #include <string.h>
 
 int init_data(struct data *data, const char *name, int nx, int ny, double dx,
-              double dy, double origin_x, double origin_y, double val) {
+              double dy, double origin_x, double origin_y, float val) {
   data->name = name;
   data->nx = nx;
   data->ny = ny;
@@ -13,7 +13,7 @@ int init_data(struct data *data, const char *name, int nx, int ny, double dx,
   data->dy = dy;
   data->origin_x = origin_x;
   data->origin_y = origin_y;
-  data->values = (double *)malloc(nx * ny * sizeof(double));
+  data->values = (float *)malloc(nx * ny * sizeof(float));
   if (!data->values) {
     printf("Error: Could not allocate data\n");
     return 1;
@@ -40,7 +40,7 @@ int write_data_vtk(struct data *data, int step, int rank) {
   }
 
   uint64_t num_points = data->nx * data->ny;
-  uint64_t num_bytes = num_points * sizeof(double);
+  uint64_t num_bytes = num_points * sizeof(float);
 
   fprintf(fp,
           "<?xml version=\"1.0\"?>\n"
@@ -72,7 +72,7 @@ int write_data_vtk(struct data *data, int step, int rank) {
           data->origin_y, 0., data->nx - 1, data->ny - 1, 0, data->name);
 
   fwrite(&num_bytes, sizeof(uint64_t), 1, fp);
-  fwrite(data->values, sizeof(double), num_points, fp);
+  fwrite(data->values, sizeof(float), num_points, fp);
 
   fprintf(fp, "  </AppendedData>\n"
               "</VTKFile>\n");
